+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ New features
 * Add function ``modeling.set_initial_condition``
 * Add function ``modeling.transform_blq``
 * Add function ``modeling.get_zero_order_inputs``
+* Add function ``modeling.set_zero_order_input``
 
 Bugfixes
 ========